Fix permission guards

+++ Carbon_Vault/Carbon_Vault/Models/AuthHelper.cs
@@ -72,7 +72,7 @@
                 Subject = new ClaimsIdentity(new[]
                 {
                     new Claim(ClaimTypes.NameIdentifier, account.Id.ToString()),
-                    new Claim(ClaimTypes.Role, account.Role.ToString())
+                    new Claim(ClaimTypes.Role, ((int)account.Role).ToString())
                 }),
                 Expires = DateTime.UtcNow.AddHours(2),
                 SigningCredentials = new SigningCredentials(
